backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
pymysql

